requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.3
pandas>=2.1.3
tiktoken>=0.5.1
PyPDF2>=3.0.1
pillow>=10.1.0
tinydb>=4.8.0

